feat(Grid): Adding test to check valid value for height and width
